requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
flask-socketio==5.1.1
google-auth==2.36.0
mysql-connector-python==9.1.0
pandas==2.2.3
python-dotenv==1.0.1
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.151.0
APScheduler==3.9.1
groq==0.11.0

